+++ edx_rest_framework_extensions/__init__.py
@@ -1,3 +1,3 @@
 """ edx Django REST Framework extensions. """
 
-__version__ = '6.1.2'  # pragma: no cover
+__version__ = '6.2.0'  # pragma: no cover