+++ edx_rest_framework_extensions/middleware.py
@@ -1,6 +1,8 @@
 """
 Middleware to ensure best practices of DRF and other endpoints.
 """
+import warnings
+
 from django.utils.deprecation import MiddlewareMixin
 from edx_django_utils import monitoring
 from edx_django_utils.cache import DEFAULT_REQUEST_CACHE
@@ -9,14 +11,14 @@ from edx_rest_framework_extensions.auth.jwt.constants import USE_JWT_COOKIE_HEAD
 from edx_rest_framework_extensions.auth.jwt.cookies import jwt_cookie_name
 
 
-class RequestMetricsMiddleware(MiddlewareMixin):
+class RequestCustomAttributesMiddleware(MiddlewareMixin):
     """
-    Adds various request related metrics.
+    Adds various request related custom attributes.
 
-    Possible metrics include:
+    Possible custom attributes include:
         request_authenticated_user_set_in_middleware:
             Example values: 'process_request', 'process_view', 'process_response',
-            or 'process_exception'. Metric won't exist if user is not authenticated.
+            or 'process_exception'. Attribute won't exist if user is not authenticated.
         request_auth_type_guess: Example values include: no-user, unauthenticated,
             jwt, bearer, other-token-type, jwt-cookie, or session-or-other
             Note: These are just guesses because if a token was expired, for example,
@@ -31,7 +33,7 @@ class RequestMetricsMiddleware(MiddlewareMixin):
 
         MIDDLEWARE = (
             'edx_django_utils.cache.middleware.RequestCacheMiddleware',
-            'edx_rest_framework_extensions.middleware.RequestMetricsMiddleware',
+            'edx_rest_framework_extensions.middleware.RequestCustomAttributesMiddleware',
         )
 
     This middleware should also appear after any authentication middleware.
@@ -51,10 +53,10 @@ class RequestMetricsMiddleware(MiddlewareMixin):
 
     def process_response(self, request, response):
         """
-        Add metrics for various details of the request.
+        Add custom attributes for various details of the request.
         """
         self._cache_if_authenticated_user_found_in_middleware(request, 'process_response')
-        self._set_all_request_metrics(request)
+        self._set_all_request_attributes(request)
         return response
 
     def process_exception(self, request, exception):  # pylint: disable=unused-argument
@@ -62,57 +64,57 @@ class RequestMetricsMiddleware(MiddlewareMixin):
         Django middleware handler to process an exception
         """
         self._cache_if_authenticated_user_found_in_middleware(request, 'process_exception')
-        self._set_all_request_metrics(request)
+        self._set_all_request_attributes(request)
 
-    def _set_all_request_metrics(self, request):
+    def _set_all_request_attributes(self, request):
         """
-        Sets all the request metrics
+        Sets all the request custom attributes
         """
-        self._set_request_auth_type_guess_metric(request)
-        self._set_request_user_agent_metrics(request)
-        self._set_request_referer_metric(request)
-        self._set_request_user_id_metric(request)
-        self._set_request_authenticated_user_found_in_middleware_metric()
+        self._set_request_auth_type_guess_attribute(request)
+        self._set_request_user_agent_attributes(request)
+        self._set_request_referer_attribute(request)
+        self._set_request_user_id_attribute(request)
+        self._set_request_authenticated_user_found_in_middleware_attribute()
 
-    def _set_request_user_id_metric(self, request):
+    def _set_request_user_id_attribute(self, request):
         """
-        Add request_user_id metric
+        Add request_user_id custom attribute
 
-        Metrics:
+        Custom Attributes:
              request_user_id
         """
         if hasattr(request, 'user') and hasattr(request.user, 'id') and request.user.id:
-            monitoring.set_custom_metric('request_user_id', request.user.id)
+            monitoring.set_custom_attribute('request_user_id', request.user.id)
 
-    def _set_request_referer_metric(self, request):
+    def _set_request_referer_attribute(self, request):
         """
-        Add metric 'request_referer' for http referer.
+        Add custom attribute 'request_referer' for http referer.
         """
         if 'HTTP_REFERER' in request.META and request.META['HTTP_REFERER']:
-            monitoring.set_custom_metric('request_referer', request.META['HTTP_REFERER'])
+            monitoring.set_custom_attribute('request_referer', request.META['HTTP_REFERER'])
 
-    def _set_request_user_agent_metrics(self, request):
+    def _set_request_user_agent_attributes(self, request):
         """
-        Add metrics for user agent for python.
+        Add custom attributes for user agent for python.
 
-        Metrics:
+        Custom Attributes:
              request_user_agent
              request_client_name: The client name from edx-rest-api-client calls.
         """
         if 'HTTP_USER_AGENT' in request.META and request.META['HTTP_USER_AGENT']:
             user_agent = request.META['HTTP_USER_AGENT']
-            monitoring.set_custom_metric('request_user_agent', user_agent)
+            monitoring.set_custom_attribute('request_user_agent', user_agent)
             if user_agent:
                 # Example agent string from edx-rest-api-client:
                 #    python-requests/2.9.1 edx-rest-api-client/1.7.2 ecommerce
                 #    See https://github.com/edx/edx-rest-api-client/commit/692903c30b157f7a4edabc2f53aae1742db3a019
                 user_agent_parts = user_agent.split()
                 if len(user_agent_parts) == 3 and user_agent_parts[1].startswith('edx-rest-api-client/'):
-                    monitoring.set_custom_metric('request_client_name', user_agent_parts[2])
+                    monitoring.set_custom_attribute('request_client_name', user_agent_parts[2])
 
-    def _set_request_auth_type_guess_metric(self, request):
+    def _set_request_auth_type_guess_attribute(self, request):
         """
-        Add metric 'request_auth_type_guess' for the authentication type used.
+        Add custom attribute 'request_auth_type_guess' for the authentication type used.
 
         NOTE: This is a best guess at this point.  Possible values include:
             no-user
@@ -137,17 +139,17 @@ class RequestMetricsMiddleware(MiddlewareMixin):
             auth_type = 'jwt-cookie'
         else:
             auth_type = 'session-or-other'
-        monitoring.set_custom_metric('request_auth_type_guess', auth_type)
+        monitoring.set_custom_attribute('request_auth_type_guess', auth_type)
 
     AUTHENTICATED_USER_FOUND_CACHE_KEY = 'edx-drf-extensions.authenticated_user_found_in_middleware'
 
-    def _set_request_authenticated_user_found_in_middleware_metric(self):
+    def _set_request_authenticated_user_found_in_middleware_attribute(self):
         """
-        Add metric 'request_authenticated_user_found_in_middleware' if authenticated user was found.
+        Add custom attribute 'request_authenticated_user_found_in_middleware' if authenticated user was found.
         """
         cached_response = DEFAULT_REQUEST_CACHE.get_cached_response(self.AUTHENTICATED_USER_FOUND_CACHE_KEY)
         if cached_response.is_found:
-            monitoring.set_custom_metric(
+            monitoring.set_custom_attribute(
                 'request_authenticated_user_found_in_middleware',
                 cached_response.value
             )
@@ -164,3 +166,13 @@ class RequestMetricsMiddleware(MiddlewareMixin):
 
         if hasattr(request, 'user') and request.user and request.user.is_authenticated:
             DEFAULT_REQUEST_CACHE.set(self.AUTHENTICATED_USER_FOUND_CACHE_KEY, value)
+
+
+class RequestMetricsMiddleware(RequestCustomAttributesMiddleware):
+    """
+    Deprecated class for handling middleware. Class has been renamed to RequestCustomAttributesMiddleware.
+    """
+    def __init__(self, *args, **kwargs):
+        super(RequestMetricsMiddleware, self).__init__(*args, **kwargs)
+        msg = "Use 'RequestCustomAttributesMiddleware' in place of 'RequestMetricsMiddleware'."
+        warnings.warn(msg, DeprecationWarning)